- Move around test code - Add app-factory.spec.ts
