+++ packages/common-types/src/index.ts
@@ -1,7 +1,6 @@
 import {
   AppABIEncodings,
   AppInstanceInfo,
-  AppInstanceInstallState,
   AssetType,
   BlockchainAsset
 } from "./data-types";
@@ -20,7 +19,6 @@ import {
 export {
   AppState,
   AppInstanceID,
-  AppInstanceInstallState,
   AppAction,
   Address,
   ABIEncoding,
